server/requirements.txt
agno==1.2.6
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.1.3
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Crawl4AI==0.5.0.post8
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.5.5
fake-http-header==0.3.5
fake-useragent==2.1.0
faust-cchardet==2.1.19
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.12.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.65.0
lxml==5.3.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.2.0
multitasking==0.0.11
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.4
ollama==0.4.7
openai==1.69.0
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==10.4.0
platformdirs==4.3.7
playwright==1.51.0
primp==0.14.0
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pyee==12.1.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pypdf==5.4.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.24.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.40
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.55
youtube-transcript-api==1.0.3
zipp==3.21.0
